src/infra/http/axios-http-client.ts
import { HttpClient, HttpRequest, HttpResponse } from '@/data/protocols/http';
import axios, { AxiosResponse } from 'axios';

export class AxiosHttpClient implements HttpClient {
  async request(data: HttpRequest): Promise<HttpResponse> {
    let axiosResponse: AxiosResponse;
    const { method, url, headers, body } = data;

    try {
      axiosResponse = await axios.request({ url, method, headers, data: body });
    } catch (error) {
      axiosResponse = error.response;
    }

    return {
      statusCode: axiosResponse.status,
      body: axiosResponse.data,
    };
  }
}
